Доделать обнуление фильтра
