+++ index.html
@@ -252,8 +252,8 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="js/main.js" type="module"></script>
   <script src="nouislider/nouislider.js"></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 
 </html>
