+++ js/esc-keydown.js
@@ -1,12 +1,12 @@
-import {closeModal} from './close-modal.js';
+// import {closeModal} from './close-modal.js';
 
-const onEscKeydown = (evt) => {
-  if ((evt.key === 'Escape' ||
-    evt.key === 'Esc') &&
-    !evt.target.classList.contains('text__hashtags') &&
-    !evt.target.classList.contains('text__description')) {
-    closeModal();
-  }
-};
+// const onEscKeydown = (evt) => {
+//   if ((evt.key === 'Escape' ||
+//     evt.key === 'Esc') &&
+//     !evt.target.classList.contains('text__hashtags') &&
+//     !evt.target.classList.contains('text__description')) {
+//     closeModal();
+//   }
+// };
 
-export {onEscKeydown};
+// export {onEscKeydown};
