+++ js/listener-keydown.js
@@ -1,7 +1,7 @@
-import {onEscKeydown} from './esc-keydown.js';
+// import {onEscKeydown} from './esc-keydown.js';
 
-const listenKeydown = () => {
-  document.removeEventListener('keydown', onEscKeydown);
-};
+// const listenKeydown = () => {
+//   document.removeEventListener('keydown', onEscKeydown);
+// };
 
-export {listenKeydown};
+// export {listenKeydown};
